requirements.txt
binance-futures-connector==3.3.0